+++ src/utils/Constants.ts
@@ -19,11 +19,12 @@ const beeActions = {
   LOAD_STAGE_MODE: 'loadStageMode',
   LOAD_CONFIG: 'loadConfig',
   LOAD_ROWS: 'loadRows',
-  UPDATE_TOKEN: 'updateToken'
+  UPDATE_TOKEN: 'updateToken',
+  GET_CONFIG: 'getConfig'
 }
 
 export const mockedEmptyToken : IToken = {
-  access_token: '', status: TokenStatus.OK, v2: true
+  access_token: '', status: TokenStatus.OK, v2: true, coediting_session_id: null, shared: false
 }
 
 export default beeActions